arreglamos la base

+++ backend/main/models/planificacion_model.py
@@ -4,9 +4,9 @@ from .. import db
 class Planificacion(db.Model):
 
     __tablename__ = "planificacion"
-    
+
     id_planificacion = db.Column(db.Integer, primary_key=True)
-    horas_semanales = db.Column(db.Integer(2))
+    horas_semanales = db.Column(db.Integer)
     objetivo = db.Column(db.String(45))
     # id_clase = db.Column(db.String(45), db.ForeignKey('Clase.id_clase'))
 
@@ -19,7 +19,6 @@ class Planificacion(db.Model):
             'horas_semanales': str(self.horas_semanales),
             'objetivo': str(self.objetivo),
             # 'id_clase': self.id_clase,
-    
         }
         return planificacion_json
 
@@ -30,7 +29,7 @@ class Planificacion(db.Model):
         objetivo = planificacion_json.get('objetivo')
         # id_clase = planificacion_json.get('id_clase')
         return Planificacion(id_planificacion=id_planificacion,
-                        horas_semanales=horas_semanales,
-                        objetivo=objetivo,
-                        # id_clase=id_clase
-                        )
+                             horas_semanales=horas_semanales,
+                             objetivo=objetivo,
+                             # id_clase=id_clase
+                             )
