+++ backend/main/models/usuario_model.py
@@ -9,9 +9,9 @@ class Usuario(db.Model):
     nombre = db.Column(db.String(45), nullable=False)
     apellido = db.Column(db.String(45), nullable=False)
     direccion = db.Column(db.String(45), nullable=False)
-    edad = db.Column(db.Integer(3), nullable=False)
-    telefono = db.Column(db.Integer(15), nullable=False)
-    dni = db.Column(db.Integer(8), nullable=False)
+    edad = db.Column(db.Integer, nullable=False)
+    telefono = db.Column(db.Integer, nullable=False)
+    dni = db.Column(db.Integer, nullable=False)
     rol = db.Column(db.String(45), nullable=False)
     sexo = db.Column(db.String(2), nullable=False)
     #id_login = db.Column(db.Integer, db.ForeignKey('Login.id_login'))
@@ -47,13 +47,13 @@ class Usuario(db.Model):
         # id_login = usuario_json.get('id_login')
 
         return Usuario(id_usuario=id_usuario,
-                        nombre=nombre,
-                        apellido=apellido,
-                        direccion=direccion,
-                        edad=edad,
-                        telefono=telefono,
-                        dni=dni,
-                        rol=rol,
-                        sexo=sexo,
-                        # id_login=id_login
-                        )
+                       nombre=nombre,
+                       apellido=apellido,
+                       direccion=direccion,
+                       edad=edad,
+                       telefono=telefono,
+                       dni=dni,
+                       rol=rol,
+                       sexo=sexo,
+                       # id_login=id_login
+                       )
